move dict json file to api
