+++ src/control-plane/backend/lambda/api/config/dictionary.json
