+++ src/control-plane/backend/lambda/api/test/api/ddb-mock.ts
@@ -31,9 +31,9 @@ import { GetSecretValueCommand } from '@aws-sdk/client-secrets-manager';
 import { StartExecutionCommand } from '@aws-sdk/client-sfn';
 import { DynamoDBDocumentClient, GetCommand, GetCommandInput, PutCommand, PutCommandOutput, QueryCommand, QueryCommandInput } from '@aws-sdk/lib-dynamodb';
 import { AwsClientStub } from 'aws-sdk-client-mock';
-import dictionary from '../../../../config/dictionary.json';
 import { analyticsMetadataTable, clickStreamTableName, dictionaryTableName, prefixTimeGSIName } from '../../common/constants';
 import { IUserRole, ProjectEnvironment } from '../../common/types';
+import dictionary from '../../config/dictionary.json';
 import { IPipeline } from '../../model/pipeline';
 
 const MOCK_TOKEN = '0000-0000';
