+++ src/control-plane/backend/lambda/batch-insert-ddb/index.ts
@@ -29,7 +29,7 @@ import {
   Context,
 } from 'aws-lambda';
 import { logger } from '../../../../common/powertools';
-import dictionary from '../../config/dictionary.json';
+import dictionary from '../../lambda/api/config/dictionary.json';
 import { aws_sdk_client_common_config } from '../api/common/sdk-client-config-ln';
 
 const ddbClient = new DynamoDBClient({
